Add Black76 benchmarking
